requirements.txt
python-telegram-bot==21.10
pymysql==1.1.1
python-dotenv==1.0.1
requests~=2.32.3
numpy~=2.2.2
matplotlib~=3.10.0
aiofiles~=24.1.0
chardet~=5.2.0
pypinyin~=0.53.0